# Improve Encode performance