This is the same idea from #5 applied to the `Encode` function: Do less `*big.Int` calculations by doing `int64` ones instead. Seems to work even better here; according to the included benchmarks the improvement is between 300% and 800%.

I have also removed `bigZero` and am using the `big.Int.Sign` method instead, to see if a number is zero. I think this should also be a little more performant.

Before:
```
$ go test -bench='Encode*'
goos: openbsd
goarch: amd64
pkg: github.com/akamensky/base58
cpu: Intel(R) Core(TM) i5-5300U CPU @ 2.30GHz
BenchmarkEncode32-2       248206              4804 ns/op           6.66 MB/s
BenchmarkEncode64-2       102597             11704 ns/op           5.47 MB/s
BenchmarkEncode128-2       37663             33468 ns/op           3.82 MB/s
BenchmarkEncode256-2       10000            103079 ns/op           2.48 MB/s
BenchmarkEncode512-2        3618            342408 ns/op           1.50 MB/s
PASS
ok      github.com/akamensky/base58     7.430s
```

After:
```
$ go test -bench='Encode*'
goos: openbsd
goarch: amd64
pkg: github.com/akamensky/base58
cpu: Intel(R) Core(TM) i5-5300U CPU @ 2.30GHz
BenchmarkEncode32-2      1000000              1127 ns/op          28.39 MB/s
BenchmarkEncode64-2       632295              1975 ns/op          32.41 MB/s
BenchmarkEncode128-2      264196              4542 ns/op          28.18 MB/s
BenchmarkEncode256-2       97514             12239 ns/op          20.92 MB/s
BenchmarkEncode512-2       29928             47047 ns/op          10.88 MB/s
PASS
ok      github.com/akamensky/base58     12.729s
```